Changed datatype.

+++ Section7/BankAccount.java
@@ -18,7 +18,7 @@ public class BankAccount {
         BankAccount.deposit = deposit;
     }
     //Creating a constructor to represent the act of checking your balance.
-    public float balance()
+    public double balance()
     {
         //Returning the current balance.
         return balance;
